Implement duck shaping on SymInts
Duck shaping says that when two input tensors have the same
size, we assume they are symbolically related.  This follows
the same optimization done by inductor.

This optimization is not done completely because we don't
currently install guards corresponding to the duck shape
relationships we created, but overall the guard propagation
for dynamic shape tracing is incomplete at the moment.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 78c4e8fcc33932f6f67394170a4447a2b5cafdcc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85808