Re-introduce the functorch docs build
We deleted it when merging functorch into pytorch. This PR makes a new
functorch docs build.

The docs are relatively simple:
- cd into `functorch/docs` and run `make html` to build the docs.
- docs should get pushed to the pytorch/functorch repo's gh-pages
branch.

The long term plan is:
- one day, the functorch APIs will just be torch.* APIs, at which point
we can fold all of the functorch docs into the regular PyTorch docs
- When that happens, the functorch examples and tutorials (that are on
the functorch docs site) can be moved to the pytorch examples and
pytorch tutorials.

Test Plan:
- check docs preview
- watch this PR after it goes in

ghstack-source-id: 7feef2d3de9d48a4091fb4d4d0131d20bdb28872
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85838